Move all but a small shim in override into _distutils_hack
